Create PacienteRepository & edit ClienteRepository
Create PacienteRepository & added prop pacienteRepo to ClienteRepository
